refactor(contracts): centralize model selection schemas
