+++ apps/server/src/git/Layers/GitManager.test.ts
@@ -6,7 +6,11 @@ import * as NodeServices from "@effect/platform-node/NodeServices";
 import { it } from "@effect/vitest";
 import { Effect, FileSystem, Layer, PlatformError, Scope } from "effect";
 import { expect, vi } from "vitest";
-import type { GitActionProgressEvent, ModelSelection, ProjectRemoteTarget } from "@t3tools/contracts";
+import type {
+  GitActionProgressEvent,
+  ModelSelection,
+  ProjectRemoteTarget,
+} from "@t3tools/contracts";
 
 import { GitCommandError, GitHubCliError, TextGenerationError } from "../Errors.ts";
 import { type GitManagerShape } from "../Services/GitManager.ts";
