+++ packages/contracts/src/provider.ts
@@ -1,5 +1,6 @@
 import { Schema } from "effect";
 import { TrimmedNonEmptyString } from "./baseSchemas";
+import { ModelSelection } from "./model";
 import {
   ApprovalRequestId,
   EventId,
@@ -10,7 +11,6 @@ import {
 } from "./baseSchemas";
 import {
   ChatAttachment,
-  ModelSelection,
   PROVIDER_SEND_TURN_MAX_ATTACHMENTS,
   PROVIDER_SEND_TURN_MAX_INPUT_CHARS,
   ProviderApprovalDecision,
